backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.4.0
pyembroidery==1.4.36
numpy==1.26.3
opencv-python==4.10.0.84
scipy==1.11.4
openai==1.12.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
stripe==7.0.0
scikit-learn==1.3.2
scikit-image==0.23.2
networkx==3.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.11.0
google-auth==2.38.0
